requirements.txt
django>=5.0,<6.0
python-chess>=1.999
PyMySQL>=1.1
